each class is a binder
Most players store a class in a binder. Even like me, I have a mono-binder that store all my classes, but still, my main division are the classes.

Actually, the binders are the classes. With one regrouping neutral, multi, and weaknesses.
This is not configurable.
The number of binders is hardcoded by this strategy.

Not yet finished, each binder shows all the cards.